I have been informed by James that the cache size default is 100, not 1000. supposedly it was 1000 in TL essentials, but is not in EclipseLink

+++ jpa/plugins/org.eclipse.jpt.eclipselink.core/src/org/eclipse/jpt/eclipselink/core/internal/context/caching/Caching.java
@@ -32,7 +32,7 @@ public interface Caching extends PersistenceUnitProperties
 			static final String CACHE_SIZE_DEFAULT_PROPERTY = "cacheSizeDefaultProperty";
 			// EclipseLink key string
 			static final String ECLIPSELINK_CACHE_SIZE_DEFAULT = "eclipselink.cache.size.default";
-			static final Integer DEFAULT_CACHE_SIZE_DEFAULT = new Integer(1000);
+			static final Integer DEFAULT_CACHE_SIZE_DEFAULT = new Integer(100);
 
 	Boolean getDefaultSharedCacheDefault();
 	Boolean getSharedCacheDefault();
@@ -57,7 +57,7 @@ public interface Caching extends PersistenceUnitProperties
 			static final String CACHE_SIZE_PROPERTY = "cacheSizeProperty";
 			// EclipseLink key string
 			static final String ECLIPSELINK_CACHE_SIZE = "eclipselink.cache.size.";
-			static final Integer DEFAULT_CACHE_SIZE = new Integer(1000);
+			static final Integer DEFAULT_CACHE_SIZE = new Integer(100);
 
 	Boolean getDefaultSharedCache();
 	Boolean getSharedCache(String entityName);
